Update recipe via conda-skeleton command with all extras
command:
conda-skeleton pypi pyspark --version 2.4.0 --extra-specs pypandoc --all-extras

This also brings in pyarrow, which is one of the great features for python
that came with Spark 2.3. It comes via the "--extra-specs" like also numpy
and pandas.

It also enforces py4j==0.10.7, which akes sense to avoid serialization
errors as spark 2.4.0 includes the py4j-0.10.7.jar file (spark on master
branch is already using py4j==0.10.8.1). py4j==0.10.7 has no official
python 3.7 support, but the conda-forge py4j==0.10.7 package is noarch and
available for all python versions. This is kind of a tradeoff, sticking to
the exact py4j version defined by spark or using a unsupported one that
works officially with python 3.7.